feat(quantize): imatrix-guided quantization + pluggable strategy refactor
Import llama.cpp imatrix GGUF files via `--imatrix <path>` to improve
MFP4 quantization quality. Refactor hipfire-quantize into pluggable
strategy traits with separated format codec modules.

## Imatrix results (Qwen3.5-9B, eval_hipfire KLD, BF16 ref)

MFP4 at matched file size, no promotion differences:
- KLD vs BF16: 0.989 -> 0.799 (-19.2%)
- PPL: 21.20 -> 18.18 (-14.2%)

MQ4/HFQ4: zero effect (byte-identical output). hipfire's single
scale+zero per group has insufficient degrees of freedom for
importance weighting to diverge from min/max. MFP4 benefits because
its multi-level scale hierarchy (row FP16 + block UE8M0 + E2M1
codebook) allows clipping low-importance outliers.

## Strategy refactor

- ScaleStrategy trait: MinMaxScale (default), ImatrixFp4Scale (MFP4)
- PromotionStrategy trait: KmapPromotion (default), ImatrixPromotion
- formats/ module tree: 9 codec files extracted from main.rs
- main.rs: 4255 -> 1929 lines (CLI + orchestration only)
- Dead code deleted: weighted_affine_solve (no-op on all formats)

## Imatrix parser

- GGUF format with general.type="imatrix" detection
- Tensor name remapping: dense, DeltaNet (attn_qkv, ssm_*), MoE
  (gate/up fusion into gate_up_proj), shared expert, router
- model.language_model. prefix normalization via ImatrixData::lookup()
- Coverage validation (80% threshold warning)

## Bug fix (pre-existing)

HFQ loader resolve_idx() tries model. and model.language_model.
prefix variants, fixing loading of models quantized from safetensors
for Qwen3.5 architecture.
