+++ crates/hipfire-quantize/src/formats/fwht.rs
@@ -1,0 +1,31 @@
+/// CPU-side FWHT (Walsh-Hadamard Transform) on a 256-element group.
+/// Matches the GPU-side fwht_forward_256 in turbo_common: signs1 -> butterfly -> scale -> signs2.
+pub fn cpu_fwht_256(x: &mut [f32], signs1: &[f32], signs2: &[f32]) {
+    assert!(x.len() == 256);
+    for i in 0..256 { x[i] *= signs1[i]; }
+    let mut stride = 1;
+    while stride < 256 {
+        let mut i = 0;
+        while i < 256 {
+            for j in 0..stride {
+                let a = x[i + j];
+                let b = x[i + j + stride];
+                x[i + j] = a + b;
+                x[i + j + stride] = a - b;
+            }
+            i += stride * 2;
+        }
+        stride <<= 1;
+    }
+    let scale = 0.0625; // 1/sqrt(256) = 1/16
+    for i in 0..256 { x[i] *= scale * signs2[i]; }
+}
+
+/// Generate FWHT sign table (matches engine's gen_fwht_signs).
+pub fn gen_fwht_signs(seed: u32, n: usize) -> Vec<f32> {
+    let mut state = seed;
+    (0..n).map(|_| {
+        state = state.wrapping_mul(1103515245).wrapping_add(12345) & 0x7fffffff;
+        if (state >> 16) & 1 == 1 { 1.0f32 } else { -1.0f32 }
+    }).collect()
+}
