+++ crates/hipfire-quantize/src/formats/hfp4.rs
@@ -1,0 +1,322 @@
+//! HFP4G32 / MFP4G32 codec: E2M1 sign-magnitude FP4 with UE8M0 block
+//! exponents and FP16 row scale.
+//!
+//! Spec: docs/quant-formats/hfp4.md
+//!
+//! Per-row layout: 16-B header (row_scale_a:f16, row_scale_b:f16, block_count:u16, flags:u8, ...)
+//!                 followed by (K/32) blocks x 17 B (UE8M0:u8 + 16 B nibbles).
+//! Per element:    value = row_scale_a * 2^(block_e - 127) * E2M1_LUT[nibble]
+
+use crate::strategy::ScaleStrategy;
+use crate::hfq_writer::{f32_to_f16, f16_to_f32};
+use crate::formats::fwht::cpu_fwht_256;
+
+/// OCP E2M1 magnitude lattice (signed 4-bit FP). 16 codes: {+/-0, +/-0.5, +/-1, +/-1.5, +/-2, +/-3, +/-4, +/-6}.
+/// Order: positive 0..7, then negative 0..7 (mirrors hardware-canonical sign-magnitude packing).
+pub const E2M1_LUT: [f32; 16] = [
+    0.0, 0.5, 1.0, 1.5, 2.0, 3.0, 4.0, 6.0,
+    -0.0, -0.5, -1.0, -1.5, -2.0, -3.0, -4.0, -6.0,
+];
+
+/// E2M1 round-to-nearest in the 16-code lattice. Returns the nibble (0..15).
+/// Ties broken away from zero (consistent with FP rounding).
+pub fn e2m1_round(x: f32) -> u8 {
+    let mut best_idx = 0u8;
+    let mut best_err = f32::INFINITY;
+    for (i, &code) in E2M1_LUT.iter().enumerate() {
+        let err = (code - x).abs();
+        // Strict < ensures consistent tie-breaking by code-table order.
+        // The lattice has +0 at index 0 and -0 at index 8; +0 wins ties at zero.
+        if err < best_err {
+            best_err = err;
+            best_idx = i as u8;
+        }
+    }
+    best_idx
+}
+
+/// Quantize one row of K FP32 weights to HFP4G32 byte format.
+///
+/// K must be a multiple of 32 (hipfire model dims always satisfy this).
+/// Returns 16-B header + (K/32) x 17-B blocks = 16 + 17 * (K/32) bytes.
+pub fn quantize_hfp4g32_row(row: &[f32], strategy: &dyn ScaleStrategy) -> Vec<u8> {
+    assert!(row.len() % 32 == 0, "HFP4G32 requires K%32 == 0, got K={}", row.len());
+    let k = row.len();
+    let n_blocks = k / 32;
+    let row_bytes = 16 + n_blocks * 17;
+    let mut out = vec![0u8; row_bytes];
+
+    // ── Row-level scale selection ────────────────────────────────────────
+    let row_scale_a = strategy.solve_fp4_row_scale(row);
+    let inv_row_scale = if row_scale_a > 0.0 { 1.0 / row_scale_a } else { 0.0 };
+
+    // Header.
+    out[0..2].copy_from_slice(&f32_to_f16(row_scale_a).to_le_bytes());
+    out[2..4].copy_from_slice(&0u16.to_le_bytes());           // row_scale_b unused in v1
+    out[4..6].copy_from_slice(&(n_blocks as u16).to_le_bytes()); // block_count
+    out[6] = 0u8;                                              // format_flags = 0 (no rotation)
+    out[7] = 0u8;                                              // reserved
+
+    // Per-block payload.
+    for b in 0..n_blocks {
+        let block_start = b * 32;
+        let block = &row[block_start..block_start + 32];
+
+        // Pick block exponent.
+        let block_max_abs = block.iter().cloned().fold(0.0f32, |m, v| m.max(v.abs()));
+        let block_max_normalized = block_max_abs * inv_row_scale;
+
+        let default_e: u8 = if block_max_normalized > 0.0 {
+            let log_ratio = (block_max_normalized / 6.0).log2();
+            (log_ratio.ceil() as i32 + 127).clamp(0, 254) as u8
+        } else {
+            0u8
+        };
+
+        let block_e = strategy.solve_fp4_block_e(block, inv_row_scale, default_e, b);
+
+        let block_scale_factor = ((block_e as i32 - 127) as f32).exp2();
+        let inv_block_scale = if block_scale_factor > 0.0 { 1.0 / block_scale_factor } else { 0.0 };
+
+        let payload_off = 16 + b * 17;
+        out[payload_off] = block_e;
+
+        for i in 0..16 {
+            let lo = block[2 * i] * inv_row_scale * inv_block_scale;
+            let hi = block[2 * i + 1] * inv_row_scale * inv_block_scale;
+            let lo_nibble = e2m1_round(lo);
+            let hi_nibble = e2m1_round(hi);
+            out[payload_off + 1 + i] = (lo_nibble & 0x0F) | ((hi_nibble & 0x0F) << 4);
+        }
+    }
+
+    out
+}
+
+/// Quantize a row-major 2D weight tensor of shape `[m, k]` to HFP4G32.
+/// Returns `m * (16 + 17 * (k/32))` bytes — 16-B row header + per-block payloads, repeated per row.
+///
+/// K%256 — not K%32 — because the v1 GEMV kernel
+/// (`crates/rdna-compute/src/dispatch.rs::gemv_hfp4g32`) iterates 256 elements
+/// per work-item and panics on K%256!=0. The byte format itself is K%32-aligned;
+/// the K%256 limit is a kernel-side constraint that v2 will lift. Refusing here
+/// makes the failure mode "quantize rejects bad input" rather than "runtime
+/// panics on first dispatch with a tensor a previous step already accepted."
+pub fn quantize_hfp4g32_2d(f32_data: &[f32], m: usize, k: usize, strategy: &dyn ScaleStrategy) -> Vec<u8> {
+    assert_eq!(f32_data.len(), m * k, "2D shape mismatch: {} vs {}*{}", f32_data.len(), m, k);
+    assert!(k % 256 == 0, "HFP4G32 v1 requires K%256==0 (gemv_hfp4g32 kernel constraint; v2 will lift to K%32==0), got K={}", k);
+    let row_bytes = 16 + 17 * (k / 32);
+    let mut out = Vec::with_capacity(m * row_bytes);
+    // imatrix is per-column (one value per K dim, shared across all rows). Pass the
+    // same strategy to every row.
+    for r in 0..m {
+        let row = &f32_data[r * k..(r + 1) * k];
+        out.extend_from_slice(&quantize_hfp4g32_row(row, strategy));
+    }
+    out
+}
+
+/// MFP4G32 = HFP4G32 + offline FWHT rotation. Drop-in MQ4 replacement.
+///
+/// Applies the same per-256-element FWHT as `cpu_fwht_256` (used by MQ4) to the
+/// weight matrix before HFP4G32 quantization. Runtime path applies the same
+/// FWHT to activations via `mq_rotate_x`, so `dot(rot(W), rot(x)) == dot(W, x)`
+/// (the FWHT is orthogonal). K must be a multiple of LCM(32, 256) = 256.
+///
+/// Sets per-row `format_flags` to `0x05` (bit 0 = rotation present, bits 2-3 = 01
+/// = offline FWHT). This is metadata only — the kernel can still consume the
+/// row as plain HFP4G32 because the rotation is baked into the codes.
+pub fn quantize_mfp4g32_2d(f32_data: &[f32], m: usize, k: usize, signs1: &[f32], signs2: &[f32], strategy: &dyn ScaleStrategy) -> Vec<u8> {
+    assert_eq!(f32_data.len(), m * k, "2D shape mismatch: {} vs {}*{}", f32_data.len(), m, k);
+    assert!(k % 256 == 0, "MFP4G32 requires k % 256 == 0 for 256-element FWHT, got k={}", k);
+    let row_bytes = 16 + 17 * (k / 32);
+    let mut out = Vec::with_capacity(m * row_bytes);
+
+    // Rotate one row's worth of weights in-place per 256-element segment, then
+    // quantize as HFP4G32 and stamp the rotation flag. Reuses signs1/signs2
+    // from the same `gen_fwht_signs(42, 256)` / `gen_fwht_signs(1042, 256)`
+    // pair MQ4 ships with so the runtime's mq_rotate_x undoes this rotation.
+    let mut row_buf = vec![0.0f32; k];
+    for r in 0..m {
+        row_buf.copy_from_slice(&f32_data[r * k..(r + 1) * k]);
+        // Apply 256-element FWHT to each segment of the row.
+        for seg in 0..(k / 256) {
+            cpu_fwht_256(&mut row_buf[seg * 256..(seg + 1) * 256], signs1, signs2);
+        }
+        let mut row_packed = quantize_hfp4g32_row(&row_buf, strategy);
+        // Stamp format_flags = 0x05 (bit 0 set + bits 2-3 = 01 = offline FWHT).
+        row_packed[6] = 0x05;
+        out.extend_from_slice(&row_packed);
+    }
+    out
+}
+
+/// CPU reference dequantization for HFP4G32 — bit-exact mirror of `gemv_hfp4g32.hip`'s dequant.
+/// Returns the K reconstructed FP32 weights for one row.
+#[allow(dead_code)] // used by tests + future round-trip diagnostics
+pub fn dequant_hfp4g32_row(packed: &[u8], k: usize) -> Vec<f32> {
+    assert!(k % 32 == 0, "HFP4G32 requires K%32 == 0");
+    let n_blocks = k / 32;
+    assert_eq!(packed.len(), 16 + n_blocks * 17, "HFP4G32 row size mismatch");
+
+    let row_scale_a_bits = u16::from_le_bytes([packed[0], packed[1]]);
+    let row_scale_a = f16_to_f32(row_scale_a_bits);
+
+    let mut out = vec![0.0f32; k];
+    for b in 0..n_blocks {
+        let payload_off = 16 + b * 17;
+        let block_e = packed[payload_off] as i32;
+        let block_scale = (block_e - 127) as f32;
+        let block_scale_factor = block_scale.exp2();
+        let scale = row_scale_a * block_scale_factor;
+
+        for i in 0..16 {
+            let byte = packed[payload_off + 1 + i];
+            let lo_nibble = (byte & 0x0F) as usize;
+            let hi_nibble = ((byte >> 4) & 0x0F) as usize;
+            out[b * 32 + 2 * i]     = scale * E2M1_LUT[lo_nibble];
+            out[b * 32 + 2 * i + 1] = scale * E2M1_LUT[hi_nibble];
+        }
+    }
+    out
+}
+
+#[cfg(test)]
+mod hfp4_tests {
+    use super::*;
+    use crate::strategy::MinMaxScale;
+    use crate::formats::fwht::gen_fwht_signs;
+
+    #[test]
+    fn e2m1_round_matches_lattice() {
+        // Each lattice value should round to its own code.
+        for (i, &val) in E2M1_LUT.iter().enumerate() {
+            let nibble = e2m1_round(val);
+            // +0 and -0 are both at value 0.0; either nibble is acceptable.
+            if val.abs() < 1e-6 {
+                assert!(nibble == 0 || nibble == 8, "zero rounds to nibble {}", nibble);
+            } else {
+                assert_eq!(nibble, i as u8, "code {} rounded to nibble {} not {}", i, nibble, i);
+            }
+        }
+    }
+
+    #[test]
+    fn e2m1_round_midpoint() {
+        // Halfway between +1.0 and +1.5 -> either is acceptable (tie).
+        let n = e2m1_round(1.25);
+        assert!(n == 2 || n == 3, "midpoint rounded to {}", n);
+        // Halfway between +4.0 and +6.0 (= 5.0) -> either is acceptable.
+        let n = e2m1_round(5.0);
+        assert!(n == 6 || n == 7, "5.0 rounded to {}", n);
+    }
+
+    #[test]
+    fn round_trip_constant_row() {
+        // All-1.0 row: row_scale_a = 1/6, every block_e ~ 127 + log2(1) = 127, every nibble = 2 (=1.0).
+        let row = vec![1.0f32; 64];
+        let packed = quantize_hfp4g32_row(&row, &MinMaxScale);
+        let recovered = dequant_hfp4g32_row(&packed, 64);
+        for (i, &v) in recovered.iter().enumerate() {
+            assert!((v - 1.0).abs() < 1e-2, "elem {} recovered to {}", i, v);
+        }
+    }
+
+    #[test]
+    fn round_trip_mixed_magnitudes() {
+        // Row with mixed positive/negative E2M1 magnitudes — should round-trip exactly.
+        let row: Vec<f32> = (0..64).map(|i| {
+            let v = E2M1_LUT[i % 16];
+            v * 6.0 // scale up so row_scale_a sees max abs at 6 * 6 = 36, brings code lattice back to [-6, 6]
+        }).collect();
+        let packed = quantize_hfp4g32_row(&row, &MinMaxScale);
+        let recovered = dequant_hfp4g32_row(&packed, 64);
+        // Bound: |recovered - input| <= row_scale * 2^(block_e - 127) * 0.5 (half min E2M1 step).
+        for (i, (&got, &want)) in recovered.iter().zip(row.iter()).enumerate() {
+            let rel_err = (got - want).abs() / want.abs().max(1.0);
+            assert!(rel_err < 0.1, "elem {}: got {} want {} rel_err {}", i, got, want, rel_err);
+        }
+    }
+
+    #[test]
+    fn round_trip_per_block_error_bound() {
+        // Mathematical guarantee: for every element, |recovered - original| must be <=
+        //   row_scale_a * 2^(block_e - 127) * (max_E2M1_step / 2)
+        // = effective_block_scale * 1.0  (max E2M1 step is 2.0, half = 1.0)
+        let mut rng_state: u64 = 0xdead_beef_dead_beef;
+        let mut next_uniform = || -> f32 {
+            rng_state ^= rng_state << 13;
+            rng_state ^= rng_state >> 7;
+            rng_state ^= rng_state << 17;
+            ((rng_state & 0x00FF_FFFF) as f32 / 0x0100_0000 as f32).max(1e-7)
+        };
+        // Box-Muller Gaussian std=0.5.
+        let row: Vec<f32> = (0..512).flat_map(|_| {
+            let u1 = next_uniform();
+            let u2 = next_uniform();
+            let r = (-2.0 * u1.ln()).sqrt();
+            let t = 2.0 * std::f32::consts::PI * u2;
+            [r * t.cos() * 0.5, r * t.sin() * 0.5]
+        }).collect();
+
+        let k = row.len();
+        let packed = quantize_hfp4g32_row(&row, &MinMaxScale);
+        let recovered = dequant_hfp4g32_row(&packed, k);
+
+        let row_scale_a = f16_to_f32(u16::from_le_bytes([packed[0], packed[1]]));
+
+        // Per-block half-max-step bound. Allow 1% slack for FP16 row-scale rounding.
+        for b in 0..(k / 32) {
+            let payload_off = 16 + b * 17;
+            let block_e = packed[payload_off] as i32;
+            let block_scale = ((block_e - 127) as f32).exp2();
+            let bound = row_scale_a * block_scale * 1.0 * 1.01 + 1e-5;
+            for i in 0..32 {
+                let idx = b * 32 + i;
+                let err = (recovered[idx] - row[idx]).abs();
+                assert!(err <= bound,
+                        "block {} elem {} err {} exceeds bound {} (block_e={}, row_scale_a={}, block_scale={})",
+                        b, i, err, bound, block_e, row_scale_a, block_scale);
+            }
+        }
+    }
+
+    #[test]
+    fn header_layout_matches_spec() {
+        // 64 elements = 2 blocks. Row size: 16 + 2*17 = 50 bytes.
+        let row = vec![3.0f32; 64];
+        let packed = quantize_hfp4g32_row(&row, &MinMaxScale);
+        assert_eq!(packed.len(), 50);
+        // Block count == 2.
+        let bc = u16::from_le_bytes([packed[4], packed[5]]);
+        assert_eq!(bc, 2);
+        // Format flags: rotation off, no row_scale_b.
+        assert_eq!(packed[6] & 0x0F, 0);
+        // First block UE8M0 byte at offset 16.
+        // Last block payload ends at 16 + 2*17 = 50 (= total).
+        // Sanity: row_scale_a > 0 (FP16 bits non-zero).
+        let rs_bits = u16::from_le_bytes([packed[0], packed[1]]);
+        assert_ne!(rs_bits, 0);
+    }
+
+    #[test]
+    fn mfp4_stamps_rotation_flag() {
+        // MFP4G32 must stamp format_flags = 0x05 (bit 0 + bits 2-3 = 01) in every row
+        // header so loaders/tooling can detect the offline-FWHT variant.
+        let m = 3;
+        let k = 256;
+        let signs1 = gen_fwht_signs(42, 256);
+        let signs2 = gen_fwht_signs(1042, 256);
+        let f32_data: Vec<f32> = (0..m * k).map(|i| (i as f32 * 0.001).sin()).collect();
+        let packed = quantize_mfp4g32_2d(&f32_data, m, k, &signs1, &signs2, &MinMaxScale);
+        let row_bytes = 16 + 17 * (k / 32);
+        assert_eq!(packed.len(), m * row_bytes, "MFP4G32 byte length mismatch");
+        for r in 0..m {
+            let off = r * row_bytes;
+            assert_eq!(packed[off + 6], 0x05, "row {} format_flags expected 0x05, got {:#x}", r, packed[off + 6]);
+            // block_count must equal k/32.
+            let bc = u16::from_le_bytes([packed[off + 4], packed[off + 5]]);
+            assert_eq!(bc as usize, k / 32);
+        }
+    }
+}
