+++ crates/hipfire-quantize/src/formats/hfq_sub4.rs
@@ -1,0 +1,176 @@
+use crate::strategy::ScaleStrategy;
+
+// ─── HFQ3-G256 Quantization ────────────────────────────────────────────────
+
+/// Quantize F32 weights to HFQ3-G256: 3-bit with 256-weight groups.
+/// Block: [f32 scale][f32 zero][96B packed 3-bit] = 104 bytes per 256 weights.
+pub fn quantize_hfq3g256(f32_data: &[f32], strategy: &dyn ScaleStrategy) -> Vec<u8> {
+    let group_size = 256;
+    let block_bytes = 104; // 8 metadata + 96 packed 3-bit
+    let n = f32_data.len();
+    let n_blocks = (n + group_size - 1) / group_size;
+    let mut output = vec![0u8; n_blocks * block_bytes];
+
+    for b in 0..n_blocks {
+        let start = b * group_size;
+        let end = (start + group_size).min(n);
+        let group = &f32_data[start..end];
+
+        let (scale, zero) = strategy.solve_affine(group, 8);
+        let inv_scale = if scale > 0.0 { 1.0 / scale } else { 0.0 };
+
+        let out_off = b * block_bytes;
+        output[out_off..out_off + 4].copy_from_slice(&scale.to_le_bytes());
+        output[out_off + 4..out_off + 8].copy_from_slice(&zero.to_le_bytes());
+
+        let actual_len = end - start;
+        // Pack 256 weights as 32 chunks of 8 weights × 3 bits = 3 bytes each = 96 bytes
+        // Matches the GEMV kernel's unpack: tid * 3 byte offset, 8 weights per thread.
+        for chunk in 0..32 {
+            let ci = chunk * 8; // index into group
+            let mut q = [0u8; 8];
+            for j in 0..8 {
+                let idx = ci + j;
+                let val = if idx < actual_len { group[idx] } else { zero };
+                q[j] = ((val - zero) * inv_scale + 0.5).clamp(0.0, 7.0) as u8;
+            }
+            // Pack 8 × 3-bit into 3 bytes (little-endian bitstream)
+            let b0 = (q[0] & 7) | ((q[1] & 7) << 3) | ((q[2] & 3) << 6);
+            let b1 = ((q[2] >> 2) & 1) | ((q[3] & 7) << 1) | ((q[4] & 7) << 4) | ((q[5] & 1) << 7);
+            let b2 = ((q[5] >> 1) & 3) | ((q[6] & 7) << 2) | ((q[7] & 7) << 5);
+
+            let bo = out_off + 8 + chunk * 3;
+            output[bo] = b0;
+            output[bo + 1] = b1;
+            output[bo + 2] = b2;
+        }
+    }
+
+    output
+}
+
+// ─── HFQ3-G128 Quantization ────────────────────────────────────────────────
+
+/// Quantize F32 weights to HFQ3-G128: 3-bit with 128-weight groups (finer granularity).
+/// Block: [f32 scale][f32 zero][48B packed 3-bit] = 56 bytes per 128 weights (0.4375 B/w).
+pub fn quantize_hfq3g128(f32_data: &[f32], strategy: &dyn ScaleStrategy) -> Vec<u8> {
+    let group_size = 128;
+    let block_bytes = 56; // 8 metadata + 48 packed 3-bit
+    let n = f32_data.len();
+    let n_blocks = (n + group_size - 1) / group_size;
+    let mut output = vec![0u8; n_blocks * block_bytes];
+
+    for b in 0..n_blocks {
+        let start = b * group_size;
+        let end = (start + group_size).min(n);
+        let group = &f32_data[start..end];
+
+        let (scale, zero) = strategy.solve_affine(group, 8);
+        let inv_scale = if scale > 0.0 { 1.0 / scale } else { 0.0 };
+
+        let out_off = b * block_bytes;
+        output[out_off..out_off + 4].copy_from_slice(&scale.to_le_bytes());
+        output[out_off + 4..out_off + 8].copy_from_slice(&zero.to_le_bytes());
+
+        let actual_len = end - start;
+        // 16 chunks of 8 weights × 3 bits = 3 bytes each = 48 bytes
+        for chunk in 0..16 {
+            let ci = chunk * 8;
+            let mut q = [0u8; 8];
+            for j in 0..8 {
+                let idx = ci + j;
+                let val = if idx < actual_len { group[idx] } else { zero };
+                q[j] = ((val - zero) * inv_scale + 0.5).clamp(0.0, 7.0) as u8;
+            }
+            let b0 = (q[0] & 7) | ((q[1] & 7) << 3) | ((q[2] & 3) << 6);
+            let b1 = ((q[2] >> 2) & 1) | ((q[3] & 7) << 1) | ((q[4] & 7) << 4) | ((q[5] & 1) << 7);
+            let b2 = ((q[5] >> 1) & 3) | ((q[6] & 7) << 2) | ((q[7] & 7) << 5);
+
+            let bo = out_off + 8 + chunk * 3;
+            output[bo] = b0;
+            output[bo + 1] = b1;
+            output[bo + 2] = b2;
+        }
+    }
+
+    output
+}
+
+// ─── HFQ2-G256 Quantization ────────────────────────────────────────────────
+
+/// Quantize F32 weights to HFQ2-G256: 2-bit with 256-weight groups.
+/// Block: [f32 scale][f32 zero][64B packed 2-bit] = 72 bytes per 256 weights (0.281 B/w).
+pub fn quantize_hfq2g256(f32_data: &[f32], strategy: &dyn ScaleStrategy) -> Vec<u8> {
+    let group_size = 256;
+    let block_bytes = 72; // 8 metadata + 64 packed
+    let n = f32_data.len();
+    let n_blocks = (n + group_size - 1) / group_size;
+    let mut output = vec![0u8; n_blocks * block_bytes];
+
+    for b in 0..n_blocks {
+        let start = b * group_size;
+        let end = (start + group_size).min(n);
+        let group = &f32_data[start..end];
+
+        let (scale, zero) = strategy.solve_affine(group, 4);
+        let inv_scale = if scale > 0.0 { 1.0 / scale } else { 0.0 };
+
+        let out_off = b * block_bytes;
+        output[out_off..out_off + 4].copy_from_slice(&scale.to_le_bytes());
+        output[out_off + 4..out_off + 8].copy_from_slice(&zero.to_le_bytes());
+
+        let actual_len = end - start;
+        // Pack 256 weights into 64 bytes (4 per byte at 2-bit)
+        for i in 0..64 {
+            let mut byte_val = 0u8;
+            for j in 0..4 {
+                let idx = 4 * i + j;
+                let val = if idx < actual_len { group[idx] } else { zero };
+                let q = ((val - zero) * inv_scale + 0.5) as u8;
+                byte_val |= q.min(3) << (j * 2);
+            }
+            output[out_off + 8 + i] = byte_val;
+        }
+    }
+
+    output
+}
+
+// ─── HFQ2-G128 Quantization ────────────────────────────────────────────────
+
+/// Quantize F32 weights to HFQ2-G128: 2-bit with 128-weight groups (finer granularity).
+/// Block: [f32 scale][f32 zero][32B packed 2-bit] = 40 bytes per 128 weights (0.3125 B/w).
+pub fn quantize_hfq2g128(f32_data: &[f32], strategy: &dyn ScaleStrategy) -> Vec<u8> {
+    let group_size = 128;
+    let block_bytes = 40;
+    let n = f32_data.len();
+    let n_blocks = (n + group_size - 1) / group_size;
+    let mut output = vec![0u8; n_blocks * block_bytes];
+
+    for b in 0..n_blocks {
+        let start = b * group_size;
+        let end = (start + group_size).min(n);
+        let group = &f32_data[start..end];
+
+        let (scale, zero) = strategy.solve_affine(group, 4);
+        let inv_scale = if scale > 0.0 { 1.0 / scale } else { 0.0 };
+
+        let out_off = b * block_bytes;
+        output[out_off..out_off + 4].copy_from_slice(&scale.to_le_bytes());
+        output[out_off + 4..out_off + 8].copy_from_slice(&zero.to_le_bytes());
+
+        let actual_len = end - start;
+        for i in 0..32 {
+            let mut byte_val = 0u8;
+            for j in 0..4 {
+                let idx = 4 * i + j;
+                let val = if idx < actual_len { group[idx] } else { zero };
+                let q = ((val - zero) * inv_scale + 0.5) as u8;
+                byte_val |= q.min(3) << (j * 2);
+            }
+            output[out_off + 8 + i] = byte_val;
+        }
+    }
+
+    output
+}
