+++ crates/hipfire-quantize/src/formats/mod.rs
@@ -1,0 +1,10 @@
+pub mod fwht;
+pub mod hfp4;
+pub mod hfq4;
+pub mod hfq6;
+pub mod hfq_sub4;
+pub mod mq4;
+pub mod mq6;
+pub mod mq8;
+pub mod mq_sub4;
+pub mod q8;
