+++ crates/hipfire-quantize/src/formats/mq4.rs
@@ -1,0 +1,46 @@
+use crate::formats::fwht::cpu_fwht_256;
+use crate::strategy::ScaleStrategy;
+
+/// MagnumQuant MQ4-G256: FWHT-rotated 4-bit quantization.
+/// Same binary format as HFQ4-G256 (136 bytes/group) — the rotation is baked
+/// into the weights. The GEMV kernel rotates x instead of inverse-rotating w.
+pub fn quantize_mq4g256(
+    f32_data: &[f32],
+    signs1: &[f32],
+    signs2: &[f32],
+    strategy: &dyn ScaleStrategy,
+) -> Vec<u8> {
+    let group_size = 256;
+    let block_bytes = 136;
+    let n = f32_data.len();
+    let n_blocks = (n + group_size - 1) / group_size;
+    let mut output = vec![0u8; n_blocks * block_bytes];
+
+    for b in 0..n_blocks {
+        let start = b * group_size;
+        let end = (start + group_size).min(n);
+
+        // Copy group and pad to 256
+        let mut group = [0.0f32; 256];
+        let actual_len = end - start;
+        group[..actual_len].copy_from_slice(&f32_data[start..end]);
+
+        // Apply FWHT rotation — this equalizes outliers across the group
+        cpu_fwht_256(&mut group, signs1, signs2);
+
+        let (scale, zero) = strategy.solve_affine(&group, 16);
+        let inv_scale = if scale > 0.0 { 1.0 / scale } else { 0.0 };
+
+        let out_off = b * block_bytes;
+        output[out_off..out_off + 4].copy_from_slice(&scale.to_le_bytes());
+        output[out_off + 4..out_off + 8].copy_from_slice(&zero.to_le_bytes());
+
+        for i in 0..128 {
+            let lo_q = ((group[2 * i] - zero) * inv_scale + 0.5) as u8;
+            let hi_q = ((group[2 * i + 1] - zero) * inv_scale + 0.5) as u8;
+            output[out_off + 8 + i] = lo_q.min(15) | (hi_q.min(15) << 4);
+        }
+    }
+
+    output
+}
