+++ crates/hipfire-quantize/src/formats/mq6.rs
@@ -1,0 +1,57 @@
+use crate::formats::fwht::cpu_fwht_256;
+use crate::strategy::ScaleStrategy;
+
+/// MagnumQuant MQ6-G256: FWHT-rotated 6-bit quantization.
+/// Same binary format as HFQ6-G256 (200 bytes/group) — the rotation is baked
+/// into the weights. The GEMV kernel rotates x instead of inverse-rotating w.
+pub fn quantize_mq6g256(
+    f32_data: &[f32],
+    signs1: &[f32],
+    signs2: &[f32],
+    strategy: &dyn ScaleStrategy,
+) -> Vec<u8> {
+    let group_size = 256;
+    let block_bytes = 200; // 8 (scale+zero) + 192 (packed 6-bit)
+    let n = f32_data.len();
+    let n_blocks = (n + group_size - 1) / group_size;
+    let mut output = vec![0u8; n_blocks * block_bytes];
+
+    for b in 0..n_blocks {
+        let start = b * group_size;
+        let end = (start + group_size).min(n);
+
+        // Copy group and pad to 256
+        let mut group = [0.0f32; 256];
+        let actual_len = end - start;
+        group[..actual_len].copy_from_slice(&f32_data[start..end]);
+
+        // Apply FWHT rotation — this equalizes outliers across the group
+        cpu_fwht_256(&mut group, signs1, signs2);
+
+        let (scale, zero) = strategy.solve_affine(&group, 64);
+        let inv_scale = if scale > 0.0 { 1.0 / scale } else { 0.0 };
+
+        let out_off = b * block_bytes;
+        output[out_off..out_off + 4].copy_from_slice(&scale.to_le_bytes());
+        output[out_off + 4..out_off + 8].copy_from_slice(&zero.to_le_bytes());
+
+        // Pack 4 values per 3 bytes: v0[5:0]|v1[1:0], v1[5:2]|v2[3:0], v2[5:4]|v3[5:0]
+        for i in (0..256).step_by(4) {
+            let q0 = ((group[i] - zero) * inv_scale + 0.5) as u8;
+            let q1 = ((group[i + 1] - zero) * inv_scale + 0.5) as u8;
+            let q2 = ((group[i + 2] - zero) * inv_scale + 0.5) as u8;
+            let q3 = ((group[i + 3] - zero) * inv_scale + 0.5) as u8;
+            let q0 = q0.min(63);
+            let q1 = q1.min(63);
+            let q2 = q2.min(63);
+            let q3 = q3.min(63);
+
+            let byte_off = 8 + (i / 4) * 3;
+            output[out_off + byte_off]     = q0 | (q1 << 6);
+            output[out_off + byte_off + 1] = (q1 >> 2) | (q2 << 4);
+            output[out_off + byte_off + 2] = (q2 >> 4) | (q3 << 2);
+        }
+    }
+
+    output
+}
