+++ crates/hipfire-quantize/src/formats/mq8.rs
@@ -1,0 +1,54 @@
+use crate::formats::fwht::cpu_fwht_256;
+use crate::hfq_writer::f32_to_f16;
+use crate::strategy::ScaleStrategy;
+
+/// MagnumQuant MQ8-G256: FWHT-rotated symmetric INT8 quantization.
+/// Format: [f16 scale][int8 x 256] = 258 bytes per 256 weights (1.008 B/w).
+/// Symmetric: scale = max(abs(group)) / 127, q = round(val / scale), no zero-point.
+/// Target: dp4a (v_dot4_i32_iu8) on gfx1100 for 4x VALU throughput.
+pub fn quantize_mq8g256(
+    f32_data: &[f32],
+    signs1: &[f32],
+    signs2: &[f32],
+    _strategy: &dyn ScaleStrategy,
+) -> Vec<u8> {
+    // _strategy accepted for API consistency. MQ8 uses symmetric INT8 (no zero point);
+    // group-average importance does not change the optimal symmetric scale.
+    let group_size = 256;
+    let block_bytes = 258; // 2 (f16 scale) + 256 (int8 values)
+    let n = f32_data.len();
+    let n_blocks = (n + group_size - 1) / group_size;
+    let mut output = vec![0u8; n_blocks * block_bytes];
+
+    for b in 0..n_blocks {
+        let start = b * group_size;
+        let end = (start + group_size).min(n);
+
+        // Copy and pad to 256
+        let mut group = [0.0f32; 256];
+        let actual_len = end - start;
+        group[..actual_len].copy_from_slice(&f32_data[start..end]);
+
+        // FWHT rotation
+        cpu_fwht_256(&mut group, signs1, signs2);
+
+        // Symmetric quantization: scale = max(|val|) / 127
+        let amax = group.iter().fold(0.0f32, |m, &v| m.max(v.abs()));
+        let scale = if amax > 0.0 { amax / 127.0 } else { 1.0 };
+        let inv_scale = if amax > 0.0 { 127.0 / amax } else { 0.0 };
+
+        let out_off = b * block_bytes;
+        // Store scale as f16 (2 bytes)
+        let scale_f16 = f32_to_f16(scale);
+        output[out_off] = (scale_f16 & 0xFF) as u8;
+        output[out_off + 1] = (scale_f16 >> 8) as u8;
+
+        // Quantize to signed INT8
+        for i in 0..256 {
+            let q = (group[i] * inv_scale).round().clamp(-128.0, 127.0) as i8;
+            output[out_off + 2 + i] = q as u8;
+        }
+    }
+
+    output
+}
